+++ internal/pkg/pagination/pagination.go
@@ -2,7 +2,7 @@ package pagination
 
 import "math"
 
-func Pages(count int64, size int64) int {
+func Pages(count int, size int) int {
 	pages := int(math.Ceil(float64(count) / float64(size)))
 	if count == 0 {
 		return 1
